requirements.txt
Faker==27.0.0
markovify==0.9.4
msal==1.30.0
python-dotenv==1.0.1
Requests==2.32.3
